requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-talisman==1.0.0
anthropic==0.20.0
openai==1.6.0
python-dotenv==1.0.0
markdown==3.5
bleach==6.1.0
werkzeug==3.0.1
jinja2==3.1.2
itsdangerous==2.2.0
sqlalchemy==2.0.23
stripe==7.13.0 